requirements.txt
tensorflow>=2.8.0
flask>=2.0.0
numpy>=1.19.5
matplotlib>=3.5.0
pandas>=1.3.0
Pillow>=8.4.0
scikit-learn>=1.0.0
opencv-python-headless>=4.5.0
tensorboard>=2.8.0
